requirements.txt
fastapi~=0.116.2
uvicorn~=0.36.0
sqlalchemy~=2.0.43
psycopg2-binary~=2.9.10
passlib[bcrypt]~=1.7.4
python-dotenv

pip~=25.1.1
h11~=0.16.0
typing_extensions~=4.15.0
sniffio~=1.3.1
click~=8.3.0
starlette~=0.48.0
pydantic~=2.11.9
pydantic_core~=2.33.2
anyio~=4.10.0
greenlet~=3.2.4
annotated-types~=0.7.0

fastapi
uvicorn
torch
transformers
sentence-transformers
scikit-learn
pillow
python-multipart
pandas
numpy
requests
pyscopg2
